Equipes/Equipe 06/Codigo/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
dj-database-url==0.4.1
Django==1.10.4
django-celery==3.2.1
django-pipeline==1.6.13
djangorestframework==3.6.3
futures==3.1.1
gunicorn==19.6.0
kombu==3.0.37
psycopg2==2.7.1
pytz==2017.2
redis==2.10.5
vine==1.1.3
whitenoise==3.2
